+++ test/unit_tests/models/test_integration.py
@@ -518,6 +518,7 @@ def test_model_torch_script(model):
 
     not_working_models = [
         "BIOT",
+        "Brant",
         "Labram",
         "EEGPT",
         "SSTDPN",